+++ srcs/Algos/astar.py
@@ -1,7 +1,7 @@
 import time
 from gamestate import Gamestate
 from puzzle import Puzzle
-from search import Search
+from Algos.search import Search
 import heapq
